+++ test/unit/UserSession/KohanaUserSessionTest.php
@@ -7,13 +7,12 @@
 namespace test\unit\Ingenerator\Warden\UI\Kohana\UserSession;
 
 
-use Doctrine\Instantiator\Exception\UnexpectedValueException;
 use Ingenerator\Warden\Core\Repository\ArrayUserRepository;
 use Ingenerator\Warden\Core\Repository\UnknownUserException;
 use Ingenerator\Warden\Core\Repository\UserRepository;
+use Ingenerator\Warden\UI\Kohana\Entity\LastLoginTrackingUser;
 use Ingenerator\Warden\UI\Kohana\UserSession\KohanaUserSession;
 use test\mock\Ingenerator\Warden\Core\Entity\UserStub;
-use test\mock\Ingenerator\Warden\UI\Kohana\Session\MockSession;
 use test\unit\Ingenerator\Warden\Core\UserSession\UserSessionTest;
 
 class KohanaUserSessionTest extends UserSessionTest
@@ -39,6 +38,17 @@ class KohanaUserSessionTest extends UserSessionTest
     {
         $this->newSubject()->login(UserStub::withId(152));
         $this->assertSame(152, $this->session_driver->get('user_id'));
+    }
+
+    public function test_it_bumps_and_persists_last_login_for_login_tracking_user()
+    {
+        $user = LoginTrackingUserStub::fromArray(['id' => 152, 'login_count' => 15]);
+
+        $this->user_repository = $this->getMockBuilder(UserRepository::class)->getMock();
+        $this->user_repository->expects($this->once())->method('save')->with($user);
+        $this->newSubject()->login($user);
+        $this->assertEquals(new \DateTimeImmutable, $user->getLastLogin(), 'Should have login', 1);
+        $this->assertSame(16, $user->getLoginCount(), 'Should be called once');
     }
 
     public function test_it_clears_user_id_on_logout()
@@ -106,6 +116,29 @@ class KohanaUserSessionTest extends UserSessionTest
             $this->session_driver,
             $this->user_repository
         );
+    }
+
+}
+
+class LoginTrackingUserStub extends UserStub implements LastLoginTrackingUser
+{
+    protected $last_login;
+    protected $login_count;
+
+    public function markLoggedInAt(\DateTimeImmutable $now)
+    {
+        $this->last_login = $now;
+        $this->login_count++;
+    }
+
+    public function getLastLogin()
+    {
+        return $this->last_login;
+    }
+
+    public function getLoginCount()
+    {
+        return $this->login_count;
     }
 
 }